% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{annovateBedFormat}
\alias{annovateBedFormat}
\title{Annovating BED format file with gene symbol.}
\usage{
annovateBedFormat(x, y, annote.database)
}
\arguments{
\item{x}{A character string of file path.}

\item{y}{A character string of a file to write to.}

\item{annote.database}{A character string of \emph{refGene.txt} file path. It
can be download from \href{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database}{UCSC database ftp}.
And \emph{refGene.txt} from \emph{Annovar} database is supported, too.}
}
\description{
Annovating BED format file with gene symbol. It will create a
  new file with a column named \emph{gene} adding to input file.
}
\examples{
####### Annovating BED format file with gene symbol #######
annovateBedFormat(x = system.file("extdata", 'testSample.state', package = "kfltCNV"),
  y = 'test.anno.state',
  annote.database = system.file("extdata", 'hg19_refGene_chr10.txt', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
