% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrWriteFile.R
\name{bedtoGRange}
\alias{bedtoGRange}
\title{Read BED file and return a GenomicRanges object.}
\usage{
bedtoGRange(file, genomeSeqinfo = NULL)
}
\arguments{
\item{file}{A character string of the BED file path.}

\item{genomeSeqinfo}{A GenomeInfoDb object which contain the sequence
information. It can also be extracted from a BAM file by \pkg{Rsamtools}
package \code{seqinfo(BamFile("bam_file_path"))}.}
}
\value{
A GenomicRanges object of input BED file.
}
\description{
Read BED file and return a GenomicRanges object. Three required
  BED fields are chromosome name, start and end position.
}
\examples{
####### Read BED file and return a GenomicRanges object. #######
require(GenomicRanges)
require(Rsamtools)
gr <- GRanges(Rle(c("1", "1", "2")),
  IRanges(start = c(1 ,10000, 100), end = c(1000, 100000, 10000)))
grangetoBed(gr, 'target.bed')
bedtoGRange('target.bed')

# genome seqinfo
bamSeqinfo <- seqinfo(
  BamFile(system.file("extdata", 'testSample.bam', package = "kfltCNV")))
bedtoGRange('target.bed', genomeSeqinfo = bamSeqinfo)
}
\author{
Zhan-Ni Chen
}
