% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{callGainCNV}
\alias{callGainCNV}
\alias{callLossCNV}
\alias{callCNVGene}
\title{Call CNVs.}
\usage{
callGainCNV(file, gapwidth)

callLossCNV(file, gapwidth)

callCNVGene(file, gapwidth)
}
\arguments{
\item{file}{A character string of a log2 ratio state file (<filename>.state)
path.}

\item{gapwidth}{A non-negative integer indicates the maximum gap width
between CNV region to ignore.}
}
\value{
If find no CNVs it will retrun NULL. Or else return a data frame with
  nine columns named \emph{chr}, \emph{start}, \emph{end}, \emph{svtype},
  \emph{log2ratio}, \emph{probe}, \emph{depth}, \emph{baseline} and
  \emph{cn}.
}
\description{
Read a log2 ratio state file (<filename>.state) and call CNVs.

\code{callGainCNV()} will merge and output copy number gains.

\code{callLossCNV()} will merge and output copy number losses.

\code{callCNVGene()} will compute gene's copy number in the CNV segment. It need an input file with a column named \emph{gene}. It can be created by a log2 ratio state file (<filename>.state) using the function \code{\link{annovateBedFormat}}.
}
\details{
The result is a data frame with nine columns and a \emph{gene} column if \code{callCNVGene()} is used:

    \emph{chr} - chromosome name.

    \emph{start} - start position of CNV segment.

    \emph{end} - end position of CNV segment.

    \emph{svtype} - \emph{DUP} or \emph{DEL} representing copy number gains or losses in CNV segment.

    \emph{log2ratio} - log2 ratio of CNV segment.

    \emph{probe} - the number of bins covered by CNV segment.

    \emph{depth} - median depth of CNV segment in test sample.

    \emph{baseline} - median depth of CNV segment in baseline.

    \emph{cn} - copy number of CNV segment.

    \emph{gene} - gene symbol.
}
\examples{
####### Call CNVs #######
callGainCNV(system.file("extdata", 'testSample.state', package = "kfltCNV"), gapwidth = 0)
}
\author{
Zhan-Ni Chen
}
