% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherTools.R
\name{callGenderByBam}
\alias{callGenderByBam}
\title{Call gender from BAM file.}
\usage{
callGenderByBam(bam, lower = 20)
}
\arguments{
\item{bam}{A character string of BAM file path.}

\item{lower}{A non-negative integer. Position of which coverage is lower than
the integer will not be count.}
}
\value{
A character string, \emph{Unknow}, \emph{Female} or \emph{Male}.
}
\description{
Call gender by read depth of chromosome X and Y. Compare the
  chi-squared values obtained to infer whether the male or female assumption
  fits read depth better.
}
\examples{
####### Call gender from BAM file #######
callGenderByBam(system.file("extdata", 'testSample.bam', package ="kfltCNV"))
}
\author{
Zhan-Ni Chen
}
