% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherTools.R
\name{checkBam}
\alias{checkBam}
\title{Check BAM file.}
\usage{
checkBam(x)
}
\arguments{
\item{x}{A character string or vector of BAM File path.}
}
\description{
It will stop if BAM file is illegal.
}
\examples{
####### Check BAM file #######
checkBam(system.file("extdata", 'testSample.bam', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
