% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCovMatrix.R
\name{computeMinTotalBase}
\alias{computeMinTotalBase}
\title{Compute minimum total bases of several coverage files.}
\usage{
computeMinTotalBase(x, method = "min", use.method = TRUE)
}
\arguments{
\item{x}{A data frame or a matrix, the result of function 'mergerCovFiles'.}

\item{method}{Default \emph{min}, also can input \emph{max}, \emph{median},
or \emph{mean}.}

\item{use.method}{\code{TRUE} by default, if \code{FALSE}, return a numeric
vector of total bases.}
}
\value{
A numeric value of total bases.
}
\description{
Calculate minimum total bases of several coverage files
  (<filename>.cov). Four required fields in a coverage file are chromosome
  name, start, end, and depth of coverage.
}
\examples{
####### Compute minimum total bases of several coverage files #######
computeMinTotalBase(c(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
 system.file("extdata", 'controlSample.cov', package = "kfltCNV")))#'

# output total base
computeMinTotalBase(c(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
 system.file("extdata", 'controlSample.cov', package = "kfltCNV")), use.method = FALSE)
}
\author{
Zhan-Ni Chen
}
