% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBaseline.R
\name{createBaseline}
\alias{createBaseline}
\title{Create a baseline to call CNV.}
\usage{
createBaseline(x, path, mode = "control", total.base = NULL,
  step = 10, ifByChr = FALSE)
}
\arguments{
\item{x}{A string value or vector of coverage file path.}

\item{path}{A string value or vector of coverage file path.}

\item{mode}{A character string, \emph{control}, \emph{shuffle} or \emph{movingAverage}.}

\item{total.base}{A non-negative numeric value used in mode \emph{control}
indicates a total base to scale to.}

\item{step}{A non-negative integer used in mode \emph{movingAverage}; it
specifies the step by performing moving average.}

\item{ifByChr}{A logical object used in mode \emph{shuffle}; if \code{TRUE},
it will shuffle depth in each chromosome. If \code{FALSE} of the whole
sample.}
}
\description{
Creating a baseline by \emph{control}, \emph{shuffle} or
  \emph{movingAverage}.
}
\details{
The \emph{control} mode (by default) will compute median depth of
  control samples in each region as the baseline. More than one coverage file
  (<filename>.cov) is need to pass and if total.base is null, it will scale
  to min total base.

  The \emph{shuffle} mode will shuffle the depth of each chromosome or the
  whole sample (when \emph{isByChr} set to \code{FALSE}) to create a
  baseline. If several files input, the first one will be useing.

  The \emph{movingAverage} mode will compute moving average depth of each
  chromosome or the whole sample to create a baseline. If several files
  input, the first one will be useing.
}
\examples{
####### Create a baseline to call CNV #######
# control
createBaseline(c(system.file("extdata", 'controlSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample2.cov', package = "kfltCNV")),
  path = 'baseline.control.cov')

# shuffle
createBaseline(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  mode = "shuffle", path = 'baseline.shuffle.cov')

# movingAverage
createBaseline(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  mode = "movingAverage", path = 'baseline.movingAverage.cov')
}
\author{
Zhan-Ni Chen
}
