% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBaseline.R
\name{createControlBaseline}
\alias{createControlBaseline}
\title{Create a baseline with control samples to call CNV.}
\usage{
createControlBaseline(x)
}
\arguments{
\item{x}{A data frame or a matrix with the same format as the result of
function \code{\link{mergerCovFiles}}, of which first three columns are
chromosome, start and end, and following with several samples' depth. It
returns a four-column data frame which has the same first three columns as
input but the forth column is median depth of samples named \emph{baseline}.}
}
\value{
A data frame with four columns named \emph{chr}, \emph{start},
  \emph{end}, \emph{baseline}.
}
\description{
Compute median depth of control samples in each region as the
  baseline for calling CNV.
}
\examples{
####### Create baseline to call CNV #######
covmatrix <- mergerCovFiles(c(
  system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample2.cov', package = "kfltCNV")))

normal.coveragematrix <- normalizeCovMatrix(covmatrix)
baseline.control <- createControlBaseline(normal.coveragematrix[, -4])
}
\author{
Zhan-Ni Chen
}
