% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBaseline.R
\name{createMoveAveBaseline}
\alias{createMoveAveBaseline}
\title{Creat a baseline by moving average.}
\usage{
createMoveAveBaseline(x, step)
}
\arguments{
\item{x}{a data frame or a matrix as the same format as the result of
function \code{\link{mergerCovFiles}}, with first three columns of
chromosome, start and end position, and forth column is depth to compute
moving average. It returns a data frame of which first three columns are
the same as input but the forth one is a moving average depth named
\emph{baseline}.}

\item{step}{A non-negative integer; it specifies the step by performing
moving average.}
}
\value{
A data frame with four columns.
}
\description{
Compute moving average depth of each chromosome to create a
  baseline depth.
}
\examples{
####### Creat a baseline by moving average #######
tumor.coveragematrix <- read.table(
  system.file("extdata", 'testSample.cov', package = "kfltCNV"), header = FALSE)
colnames(tumor.coveragematrix) <- c('chr', 'start', 'end', 'testSample')
baseline.moveaverage <- createMoveAveBaseline(tumor.coveragematrix, step=10)
}
\author{
Zhan-Ni Chen
}
