% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBaseline.R
\name{createShuffleBaseline}
\alias{createShuffleBaseline}
\title{Creat a baseline by shuffling.}
\usage{
createShuffleBaseline(x, ifByChr = FALSE)
}
\arguments{
\item{x}{A data frame or a matrix as the same format as the result of
function \code{\link{mergerCovFiles}}, with first three columns of
chromosome, start and end, and forth column is depth to shuffle. It returns
a data frame of which first three columns are the same as input but the
forth one is a shuffled depth named \emph{baseline}.}

\item{ifByChr}{Logical; if \code{TRUE}, it will shuffle depth of each
chromosome. If \code{FALSE} of the whole sample.}
}
\value{
A data frame with four columns named \emph{chr}, \emph{start},
  \emph{end}, \emph{baseline}.
}
\description{
Shuffle the depth of each chromosome or the whole sample to
  create a baseline.
}
\examples{
####### Creat a baseline by shuffling #######
tumor.coveragematrix <- read.table(
  system.file("extdata", 'testSample.cov', package = "kfltCNV"), header = FALSE)
colnames(tumor.coveragematrix) <- c('chr', 'start', 'end', 'testSample')
baseline.shuffle <- createShuffleBaseline(tumor.coveragematrix)
}
\author{
Zhan-Ni Chen
}
