% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrWriteFile.R
\name{grangetoBed}
\alias{grangetoBed}
\title{Write a GenomicRanges object into a BED file.}
\usage{
grangetoBed(gr, path)
}
\arguments{
\item{gr}{A GenomicRanges object.}

\item{path}{A character string of the BED file path to write to.}
}
\description{
Write a GenomicRanges object into a BED file with three required
  BED fields, i.e. chromosome name, start and end position.
}
\examples{
####### Write a GenomicRanges object into a BED file #######
require(GenomicRanges)
gr <- GRanges(Rle(c("1", "1", "2")),
  IRanges(start = c(1 ,10000, 100), end = c(1000, 100000, 10000)))
grangetoBed(gr, 'target.bed')
}
\author{
Zhan-Ni Chen
}
