% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCovMatrix.R
\name{isSameBedFile}
\alias{isSameBedFile}
\title{Test two BED files for equal.}
\usage{
isSameBedFile(x, y)
}
\arguments{
\item{x, y}{Character strings, a BED file or a file with first three columns
the same format as a BED file, such as a coverage file (<filename>.cov).}
}
\value{
It returns \code{TRUE} if two BED file contain the same regions,
  \code{FALSE} in every other case.
}
\description{
Read two BED files and change them to sorted GenomicRanges
  objects, then use \code{identical()} to test for exactly equal. It returns
  \code{TRUE} in this case, \code{FALSE} in every other case. Three required
  BED fields are chromosome name, start and end position.
}
\examples{
####### Test two BED files for equal #######
isSameBedFile(x = system.file("extdata", 'chr10_exome.bed', package = "kfltCNV"),
y = system.file("extdata", 'testSample.cov', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
