% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfltLog2Ratio.R
\name{kflt}
\alias{kflt}
\title{Kalman filter.}
\usage{
kflt(yt, ct, model = list(p = 0, d = 1, q = 5, bin.size = 1e+05, sn =
  2.5, method = "nlminb"), nafun = na.locf, smooth = 0)
}
\arguments{
\item{yt}{A matrix containing the observations. \code{NA} values are allowed.}

\item{ct}{A matrix giving the intercept of the measurement equation.}

\item{model}{A list object contains elements named \emph{p}, \emph{d},
\emph{q}, \emph{bin.size}, \emph{sn} and \emph{method}. See details in the
function \code{\link{kflt}}. By default it is for log2 ratio data.}

\item{nafun}{A function's name to fill \code{NA} value. Default is
\code{\link[zoo]{na.locf}}.}

\item{smooth}{A non-negative integer indicating the step width to smooth the
state after Kalman filtering.}
}
\description{
This function allows for Kalman filtering a numeric vector use
  the \code{fkf()} function in the package \pkg{FKF}. See the details in the
  funtion \code{\link[FKF]{fkf}}.
}
\details{
The Kalman filter is a set of recursive linear estimation steps,
  which computes an estimation value at time \emph{'t+1'} based on the
  estimation value at time \emph{'t'} and the observation value at time
  \emph{'t+1'}. Its estimates are linear combinations of Gaussian data when
  the noise disturbances are Gaussian.

Read depth data from Next-Generation Sequencing are proved to be a
combinations of Gaussian data. Perform log-ratio transformation on read depth
to reduce system error from sequencing process, and smooth by moving-average.
So the log2 ratio data is Gaussian and can be simulated by the Kalman filter.

Based on the Box-Jenkins Approach, the Log2 ratio data fits a model of
\emph{ARMA(0,5)-process} with zero (p) autoregressive term and five (q)
moving-average terms.

Ande parameters in detail are:

    \emph{yt} - a matrix containing the observation and \code{NA} values are
    allowed. It is a \emph{1 * n} matrix with log2 ratio data of n region.

    \emph{ct} - a matrix giving the intercept of the measurement equation. It
    is a \emph{1 * 1} matrix and value is 0 indicating that log2 ratio is
    symmetrical around 0.

    \emph{model} - a list object containing six elements:

        \emph{p} - a non-negative integer providing the number of
        autoregressive terms.

        \emph{q} - a non-negative integer providing the number of
        moving-average terms.

        \emph{d} - a non-negative integer. If d is over 0, it will difference
        the data to reduce the changes in the level.

        \emph{bin.size} - a non-negative integer indicates the width of each
        window.

        \emph{sn} - a numberic value providing an average signal-to-noise
        ratio. If input data is unstable it can be turn down.

        \emph{method} - the method to optimize the model. By default it is
        \emph{nlminb} which use the function \code{\link[stats]{nlminb}}.
        Other methods are \emph{Nelder-Mead}, \emph{BFGS}, \emph{CG},
        \emph{L-BFGS-B}, \emph{SANN} and \emph{Brent} which use the function
        \code{\link[stats]{optim}}. See details in description of the
        function \code{\link[stats]{optim}}.

    \emph{nafun} - the function name in the package \pkg{zoo} to fill a
    \code{NA} data point. By default is \code{\link[zoo]{na.locf}}.

    \emph{smooth} - an non-negative integer indicates the step when do
    moving-average. If smooth is 0, it return the original state of Kalman
    filtering.
}
\examples{
####### Kalman filter #######
kflt(
yt = matrix(data = log2(rnorm(1000, mean = 2, sd = 0.1)/2), nrow = 1),
ct = matrix(data = 1, nrow = 1))
}
\author{
Qi-Yuan Li
}
