% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfltBatch.R
\name{kfltBatch}
\alias{kfltBatch}
\title{Batch pipeline of calling CNV.}
\usage{
kfltBatch(testBamFile, controlBamFile = NULL, baselineFile = NULL,
  mode = "control", bedFile = NULL, binSize = NULL, outDir = NULL,
  annote.database = NULL, thread = 1, badDepth = 10, gapwidth = 0,
  threshold = 0, min.probes = 1, is.plot = TRUE,
  is.run.by.gender = FALSE, shuffle.ifByChr = FALSE,
  moveAverage.step = 10)
}
\arguments{
\item{testBamFile}{A character string of test BAM file path.}

\item{controlBamFile}{A character string or vector of control BAM files path.}

\item{baselineFile}{A character string of baseline coverage file path.}

\item{mode}{A character string, \emph{control}, \emph{shuffle} or \emph{movingAverage}.}

\item{bedFile}{A character string of BED file path.}

\item{binSize}{A non-negative integer giving the width of each bin or window,
such as \strong{1E2}, \strong{1e2} or \strong{100}.}

\item{outDir}{A character string of directory path to write to.}

\item{annote.database}{A character string of \emph{refGene.txt} file path. It
can be download from \href{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database}{UCSC database ftp}.
And \emph{refGene.txt} from \emph{Annovar} database is supported, too.}

\item{thread}{A non-negative integer providing the number of threads.}

\item{badDepth}{A numeric value; a baseline depth low than badDepth will be
set to \code{NA}.}

\item{gapwidth}{A non-negative integer indicates the maximum gap width
between CNV region to ignore.}

\item{threshold}{A non-negative numeric value. A CNV of which log2 ratio is
out of range from inverse \emph{threshold} to \emph{threshold} will not
export.}

\item{min.probes}{A non-negative integer. A CNV of which probe is under
\emph{min.probes} will not export.}

\item{is.plot}{Logical. Whether to plot results.}

\item{is.run.by.gender}{Logial. Whether to call CNV based on sample's gender.}

\item{shuffle.ifByChr}{A logical object used in mode \emph{shuffle}; if
\code{TRUE}, it will shuffle depth of each chromosome. If \code{FALSE} of
the whole sample.}

\item{moveAverage.step}{A non-negative integer used in mode
\emph{movingAverage}; it specifies the step by performing moving average.}
}
\description{
A pipeline for using BAM files or baseline coverage files to
  call CNV by the Kalman Filter.
}
\examples{
####### Batch pipeline of calling CNV #######
testBamFile <- system.file("extdata", 'testSample.bam', package ="kfltCNV")
controlBamFile <- c(system.file("extdata", 'controlSample.bam', package ="kfltCNV"),
  system.file("extdata", 'controlSample2.bam', package ="kfltCNV"))
bedFile <- system.file("extdata", 'chr10_exome.bed', package ="kfltCNV")
binSize <- 1E2
kfltBatch(testBamFile, controlBamFile = controlBamFile, bedFile = bedFile, binSize = 1E2)
}
\author{
Zhan-Ni Chen
}
