% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCovMatrix.R
\name{linearInterpolationCov}
\alias{linearInterpolationCov}
\title{Align a coverage file.}
\usage{
linearInterpolationCov(covFile, axisFile)
}
\arguments{
\item{covFile}{A character string of a coverage file path.}

\item{axisFile}{A character string of a BED or BED-like file with first three
columns of chromosome name, start and end position.}
}
\value{
A GenomicRanges object of which region is consistent with
  \emph{axisFile} and has an updated depth column.
}
\description{
Align a coverage file (<filename>.cov) by a BED-like file and
  fit \code{NA} data point by linear interpolation.
}
\details{
Four required fields in a coverage file are chromosome name, start
  and end position, and depth of coverage. Three required BED-like fields are
  chromosome name, start and end position.
}
\examples{
####### Align a coverage file #######
require(GenomicRanges)
gr <- readCovFile(system.file("extdata", 'testSample.cov', package = "kfltCNV"))
writeCovFile(shift(gr, shift = 10L), 'shift.cov')

linearInterpolationCov(covFile = 'shift.cov',
axisFile = system.file("extdata", 'testSample.cov', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
