% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCovMatrix.R
\name{normalizeCovMatrix}
\alias{normalizeCovMatrix}
\title{Normalize a coverage matrix.}
\usage{
normalizeCovMatrix(x, total.base = NULL)
}
\arguments{
\item{x}{A data frame or a matrix, the result of function
\code{mergerCovFiles()}.}

\item{total.base}{A non-negative integer, indicates a total base to scale to.}
}
\value{
A data frame with the same format as input.
}
\description{
Normalize every sample's coverage to the same total bases, i.e.
  the minimum total coverage of input samples. Input the result of function
  \code{\link{mergerCovFiles}}, which is a data frame with first three
  columns of chromosome, start, end, and following by columns of several
  samples' depths, and each row means different regions.
}
\examples{
####### Normalize a coverage matrix #######
covmatrix <- mergerCovFiles(c(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample.cov', package = "kfltCNV")))
normalizeCovMatrix(covmatrix)
}
\author{
Zhan-Ni Chen
}
