% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{performCallCNV}
\alias{performCallCNV}
\title{Perform call CNVs.}
\usage{
performCallCNV(file, outPrefix = NULL, id = NULL, gapwidth = 500,
  annote.database = NULL, out.type = "all", threshold = 0,
  min.probes = 1)
}
\arguments{
\item{file}{A character string of a log2 ratio state file (<filename>.state)
path.}

\item{outPrefix}{A character string, indicates output files' prefix. By
default it is the prefix of input file.}

\item{id}{A character string of sample id in VCF file. By default it is the
prefix of input file.}

\item{gapwidth}{A non-negative integer indicates the maximum gap width
between CNV region to ignore.}

\item{annote.database}{A character string of \emph{refGene.txt} file path. It
can be download from \href{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database}{UCSC database ftp}.
And \emph{refGene.txt} from \emph{Annovar} database is supported, too.}

\item{out.type}{A character string, \emph{vcf}, \emph{table} or \emph{all}.
By default it will export both VCF file and tabular file.}

\item{threshold}{A non-negative numeric value. A CNV of which log2 ratio is
out of range from inverse \emph{threshold} to \emph{threshold} will not
export.}

\item{min.probes}{A non-negative integer. A CNV of which probe is under
\emph{min.probes} will not export.}
}
\description{
Perform call CNVs from log2 ratio state file (<filename>.state).
  If \emph{annote.database} isn't \code{NULL}, it will create the gene CNV
  result.
}
\details{
The results are tabular or
  \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{VCFv4.2} format
  file.

The tabular file has nine columns while a gene CNV result will also have a column named \emph{gene}.

    \emph{chr} - chromosome name.

    \emph{start} - start position of CNV segment.

    \emph{end} - end position of CNV segment.

    \emph{svtype} - \emph{DUP} or \emph{DEL} representing copy number gains or losses in CNV segment.

    \emph{log2ratio} - log2 ratio of CNV segment.

    \emph{probe} - the number of bins covered by CNV segment.

    \emph{depth} - median depth of CNV segment in test sample.

    \emph{baseline} - median depth of CNV segment in baseline.

    \emph{cn} - copy number of CNV segment.

    \emph{gene} - gene symbol.
}
\examples{
####### Perform call CNVs #######
performCallCNV(
    file =  system.file("extdata", 'testSample.state', package = "kfltCNV"),
    annote.database = system.file("extdata", 'hg19_refGene_chr10.txt', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
