% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthOfGRanges.R
\name{performCreateCovFile}
\alias{performCreateCovFile}
\title{Perform calculating depth of coverage.}
\usage{
performCreateCovFile(bamFiles, bedFile = NULL, outDir = NULL,
  width = NULL, thread = 1, batch = 1000, mapq.filter = 30)
}
\arguments{
\item{bamFiles}{A character string or vector of the BAM files path.}

\item{bedFile}{A BED file path.}

\item{outDir}{A character string of directory to output coverage files
(<filename>.cov). Default is the current folder.}

\item{width}{An integer giving the width of each window, such as
\strong{1E2}, \strong{1e2} or \strong{100}.}

\item{thread}{An integer providing the number of thread.}

\item{batch}{An integer giving how many GRanges are performed in a batch.}

\item{mapq.filter}{A non-negative integer specifying the minimum mapping
quality to include. BAM reads with mapping qualities less than mapqFilter
are discarded.}
}
\description{
Input a BAM file or files to calculate the total bases
  overlapping the region. Sequencing depth of coverage in this region equals
  to total bases divide by width of region.
}
\details{
The input BAM file need a BAM index file (<filename>.bai). The index
  file can be generated by \emph{samtools} or other BAM/SAM tools after
  sorting. Function will stop if the BAM file lose an index file or its index
  file is older than its own.

  If \emph{bedFile} is not \code{NULL}, it will be splited unless the
  \emph{width} option is also \code{NULL}. If a BED file is not provided,
  windows will be created from the whole genome, by default 100,000 bp each
  window. Coverage of depth will be calculated by each window. Three required
  BED fields are chromosome name, start and end position.

  It can perform in multithreading with number of threads you input. If the
  \emph{thread} is over 1, it will perform in multithreading, default by
  single thread. Multithreading uses \code{parLapply()} function in
  \pkg{parallel} package and it can run on Linux but not Windows system. And
  the \emph{batch} option limits the number of GRanges in a batch to run in a
  thread.

  The result are coverage files (<filename>.cov) which have four
  columns indicate chromosome name, chromosome start, chromosome end and
  depth of coverage, without a header line.
}
\examples{
####### Perform calculating depth of coverage #######
performCreateCovFile(
  bamFiles = system.file("extdata", 'testSample.bam', package = "kfltCNV"),
  bedFile = system.file("extdata", 'chr10_exome.bed', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
