% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfltLog2Ratio.R
\name{performFitCovFile}
\alias{performFitCovFile}
\title{Prepare a log2 ratio file.}
\usage{
performFitCovFile(testFile, path, baselineFile = NULL,
  controlFile = NULL, badDepth = 10)
}
\arguments{
\item{testFile}{A character string of a coverage file path to call CNV.}

\item{path}{A character string of the log2 ratio file to output.}

\item{baselineFile}{A character string of a baseline coverage file.}

\item{controlFile}{A character string or vector of control samples' coverage
files.}

\item{badDepth}{A non-negative numeric value. The region of which depth in
baseline low than badDepth will be set to \code{NA}.}
}
\description{
Prepare a log2 ratio file (<filename>.fit) to do Kalman filtering.
}
\details{
A log2 ratio file has six columns respectively named \emph{chr},
  \emph{start}, \emph{end}, \emph{log2ratio}, \emph{depth} and
  \emph{baseline}.

  And \emph{log2ratio} is the log2 ratio of test sample's depth to the
  baseline. The columns named \emph{depth} and \emph{baseline} are depth of
  coverage respectively in test sample and the baseline.

  If \emph{baselineFile} isn't \code{NULL}, it will use baseline coverage
  file to create the log2 ratio file. When you input control samples'
  coverage files, it will use control samples to create a baseline at first.
}
\examples{
####### Prepare a log2 ratio file #######
testFile <- system.file("extdata", 'testSample.cov', package = "kfltCNV")
controlFile <- c(system.file("extdata", 'controlSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample2.cov', package = "kfltCNV"))

performFitCovFile(testFile, path = 'testSample.fit', controlFile = controlFile)
}
\author{
Zhan-Ni Chen
}
