% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKfltResult.R
\name{plotCoverageUniformity}
\alias{plotCoverageUniformity}
\title{Plot the unifomity of coverage among samples.}
\usage{
plotCoverageUniformity(x)
}
\arguments{
\item{x}{A numeric matrix or a data frame.}
}
\value{
A boxplot R object.
}
\description{
Compute correlation between columns of a numeric matrix or data
  frame and plot the distribution of correlations.
}
\examples{
####### Plot the coverage unifomity of samples input #######
covmatrix <- mergerCovFiles(c(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample.cov', package = "kfltCNV"),
  system.file("extdata", 'controlSample2.cov', package = "kfltCNV")))
plotCoverageUniformity(covmatrix[,4:6])
}
\author{
Zhan-Ni Chen
}
