% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKfltResult.R
\name{plotKfltResult}
\alias{plotKfltResult}
\title{Plot kflt results.}
\usage{
plotKfltResult(file)
}
\arguments{
\item{file}{A character string of Kalman filter result file
(<fileName>.state).}
}
\description{
Plot the log2 ratio state after Kalman filtering.
}
\examples{
####### Plot kflt results #######
plotKfltResult(system.file("extdata", 'testSample.state', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
