% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrWriteFile.R
\name{readCovFile}
\alias{readCovFile}
\title{Read a coverage file.}
\usage{
readCovFile(file)
}
\arguments{
\item{file}{A character string of coverage file path.}
}
\value{
A GenomicRanges object of genome region with a column named depth.
}
\description{
Read a coverage file (<filename>.cov) and change it to a sorted
  GenomicRanges object. Four required a coverage file fields are chromosome
  name, start, end and coverage of depth.
}
\examples{
####### Read a coverage file #######
readCovFile(system.file("extdata", 'testSample.cov', package = "kfltCNV"))
}
\author{
Zhan-Ni Chen
}
