% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCovMatrix.R
\name{scaleCovFile}
\alias{scaleCovFile}
\title{Scale a coverage file to a provided total base.}
\usage{
scaleCovFile(x, path, total.base)
}
\arguments{
\item{x}{A character string of a coverage file path.}

\item{path}{A character string of a file to write to.}

\item{total.base}{A non-negative integer, indicates a total base to scale to.}
}
\description{
Scale a coverage file (<filename>.cov) by a provided total base.
  Four required fields in a coverage file are chromosome name, start and end
  position, and depth of coverage.
}
\examples{
####### Scale a coverage file to a provided total base #######
scaleCovFile(system.file("extdata", 'testSample.cov', package = "kfltCNV"),
  path = 'testSample.nor.cov', total.base = 1E6)
}
\author{
Zhan-Ni Chen
}
