% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBed.R
\name{splitGenome}
\alias{splitGenome}
\title{Spliting genome into windows.}
\usage{
splitGenome(genomeSeqinfo, width)
}
\arguments{
\item{genomeSeqinfo}{A GenomeInfoDb object which contain the sequence
information. It can also be extracted from a BAM file by \pkg{Rsamtools}
package \code{seqinfo(BamFile("bam_file_path"))}.}

\item{width}{An integer giving the width of each window, such as
\strong{1E2}, \strong{1e2} or \strong{100}.}
}
\value{
A GenomicRanges object of splited regions from genome.
}
\description{
Split genome into windows with provided width.
}
\examples{
####### Spliting genome into windows #######
require(Rsamtools)
bamSeqinfo <- seqinfo(BamFile(system.file("extdata", 'testSample.bam', package = "kfltCNV")))
splitGenome <- splitGenome(bamSeqinfo, width = 1E7)
}
\author{
Zhan-Ni Chen
}
