% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOrWriteFile.R
\name{writeCovFile}
\alias{writeCovFile}
\title{Write a GenomicRanges object into a coverage file.}
\usage{
writeCovFile(gr, path)
}
\arguments{
\item{gr}{A GenomicRanges object with a column named depth.}

\item{path}{A character string of a coverage file path to write to.}
}
\description{
Write a GenomicRanges object with a column named depth into a
  coverage file (<filename>.cov). The result file does not have a header
  line, and its four columns respectively indicate chromosome name, start
  position, end position, and depth.
}
\examples{
####### Write a GenomicRanges object into a coverage file #######
gr <- readCovFile(system.file("extdata", 'testSample.cov', package = "kfltCNV"))
writeCovFile(gr, 'sample.cov')
}
\author{
Zhan-Ni Chen
}
